+++ common/features.h
@@ -5,19 +5,17 @@
 #include <ccan/tal/tal.h>
 
 /* Returns true if we're OK with all these offered features. */
-bool features_supported(const u8 *globalfeatures, const u8 *localfeatures);
+bool features_supported(const u8 *features);
 
 /* For sending our features: tal_count() returns length. */
-u8 *get_offered_globalfeatures(const tal_t *ctx);
-u8 *get_offered_localfeatures(const tal_t *ctx);
+u8 *get_offered_features(const tal_t *ctx);
 u8 *get_offered_nodefeatures(const tal_t *ctx);
 
 /* Is this feature bit requested? (Either compulsory or optional) */
 bool feature_offered(const u8 *features, size_t f);
 
 /* Was this feature bit offered by them and us? */
-bool local_feature_negotiated(const u8 *lfeatures, size_t f);
-bool global_feature_negotiated(const u8 *gfeatures, size_t f);
+bool feature_negotiated(const u8 *lfeatures, size_t f);
 
 /* Return a list of what features we advertize. */
 const char **list_supported_features(const tal_t *ctx);
@@ -47,18 +45,14 @@ void set_feature_bit(u8 **ptr, u32 bit);
  * | 4/5  | `option_upfront_shutdown_script` |...
  * | 6/7  | `gossip_queries`                 |...
  * | 10/11 | `gossip_queries_ex`             |...
- */
-#define LOCAL_DATA_LOSS_PROTECT			0
-#define LOCAL_INITIAL_ROUTING_SYNC		2
-#define LOCAL_UPFRONT_SHUTDOWN_SCRIPT		4
-#define LOCAL_GOSSIP_QUERIES			6
-#define LOCAL_GOSSIP_QUERIES_EX			10
-
-/* BOLT #9:
- * | Bits  | Name                            |...
  * | 12/13| `option_static_remotekey`        |...
  */
-#define LOCAL_STATIC_REMOTEKEY			12
+#define OPT_DATA_LOSS_PROTECT			0
+#define OPT_INITIAL_ROUTING_SYNC		2
+#define OPT_UPFRONT_SHUTDOWN_SCRIPT		4
+#define OPT_GOSSIP_QUERIES			6
+#define OPT_GOSSIP_QUERIES_EX			10
+#define OPT_STATIC_REMOTEKEY			12
 
 /* BOLT #9:
  *
@@ -67,6 +61,6 @@ void set_feature_bit(u8 **ptr, u32 bit);
  * | Bits | Name              | ...
  * | 8/9  | `var_onion_optin` | ...
  */
-#define GLOBAL_VAR_ONION			8
+#define OPT_VAR_ONION				8
 
 #endif /* LIGHTNING_COMMON_FEATURES_H */